test: easyio and fileio test
