+++ fileio/fileio_test.go
@@ -1,13 +1,10 @@
 package fileio
 
 import (
-	"bytes"
 	"io"
 	"os"
 	"strings"
 	"testing"
-
-	"github.com/klauspost/pgzip"
 )
 
 var testfile string = "testdata/smallTest"
@@ -65,7 +62,7 @@ func TestMustCreateEmptyFilenameError(t *testing.T) {
 	MustCreate("")
 }
 
-func TestStdinSimulation(t *testing.T) {
+func TestStdin(t *testing.T) {
 	testCases := []string{
 		"Gonomics fileio stdin mock",
 		"",                 // Empty string
@@ -79,6 +76,7 @@ func TestStdinSimulation(t *testing.T) {
 			t.Fatalf("Error: Failed to create temporary file: %v", err)
 		}
 		defer mockStdin.Close()
+		defer MustRemove(mockStdin.Name())
 		if _, err := mockStdin.WriteString(input); err != nil {
 			t.Fatalf("Error: Failed to write simulated input: %v", err)
 		}
@@ -98,59 +96,6 @@ func TestStdinSimulation(t *testing.T) {
 		if reader.String() != input {
 			t.Errorf("Error: Mismatch for input %q. Expected: %q, got %q", input, input, reader.String())
 		}
-		MustRemove(mockStdin.Name())
-	}
-}
 
-func TestStdinGzip(t *testing.T) {
-	testCases := []string{
-		"fileio stdin with gzip",
-		"stdin.gz",
-	}
-
-	for _, input := range testCases {
-		// Create a temporary file to simulate stdin
-		mockStdin, err := os.CreateTemp("", "mockStdin.gz")
-		if err != nil {
-			t.Fatalf("Error: Failed to create temporary file: %v", err)
-		}
-		defer os.Remove(mockStdin.Name())
-		defer mockStdin.Close()
-
-		// Gzip the input data
-		var gzippedInput bytes.Buffer
-		gzWriter := pgzip.NewWriter(&gzippedInput)
-		if _, err := gzWriter.Write([]byte(input)); err != nil {
-			t.Fatalf("Error: Failed to gzip input: %v", err)
-		}
-		if err := gzWriter.Close(); err != nil {
-			t.Fatalf("Error: Failed to close gzip writer: %v", err)
-		}
-
-		// Write the gzipped input to the mock stdin
-		if _, err := io.Copy(mockStdin, &gzippedInput); err != nil {
-			t.Fatalf("Error: Failed to write simulated input: %v", err)
-		}
-
-		// Seek to the beginning of the mock stdin
-		if _, err := mockStdin.Seek(0, io.SeekStart); err != nil {
-			t.Fatalf("Error: Failed to seek to the start: %v", err)
-		}
-
-		// Set os.Stdin to the mock stdin
-		os.Stdin = mockStdin
-
-		// Call the function being tested
-		file := EasyOpen("/dev/stdin")
-		defer file.Close()
-
-		// Read from the file and compare with the original input
-		var reader strings.Builder
-		if _, err := io.Copy(&reader, file); err != nil {
-			t.Fatalf("Error: Failed to read from simulated stdin: %v", err)
-		}
-		if reader.String() != input {
-			t.Errorf("Error: Mismatch for input %q. Expected: %q, got %q", input, input, reader.String())
-		}
 	}
 }
